+++ tests/unit/data-table/DataTable.spec.js
@@ -1,73 +1,7 @@
-import { mount } from '@vue/test-utils'
 import DataTable from '@/components/data-table/DataTable.vue'
 
-import FfDataTableRow from '@/components/data-table/DataTableRow.vue'
-import FfDataTableCell from '@/components/data-table/DataTableCell.vue'
-import FfKebabMenu from '@/components/KebabMenu.vue'
-import FfTextInput from '@/components/form/TextInput.vue'
-import FfCheck from '@/components/Check.vue'
-
-describe('DataTable UX', () => {
-    it('supports searching rows by values', async () => {
-        const wrapper = mount(DataTable, {
-            props: {
-                columns: [{
-                    key: 'name',
-                    label: 'Name'
-                }],
-                rows: [
-                    { name: 'Apple' },
-                    { name: 'Orange', details: null },
-                    { name: 'Grapefruit', details: { color: 'Orangey' } }
-                ],
-                showSearch: true
-            },
-            global: {
-                components: {
-                    FfDataTableRow, FfDataTableCell, FfTextInput, FfCheck, FfKebabMenu
-                }
-            }
-        })
-
-        expect(wrapper.find('tbody').text()).toBe('AppleOrangeGrapefruit')
-
-        await wrapper.find('[data-form="search"] input').setValue('Orange')
-
-        expect(wrapper.find('tbody').text()).toBe('OrangeGrapefruit')
-    })
-
-    it('supports searching rows by values in specific fields only', async () => {
-        const wrapper = mount(DataTable, {
-            props: {
-                columns: [{
-                    key: 'name',
-                    label: 'Name'
-                }],
-                rows: [
-                    { name: 'Apple' },
-                    { name: 'Orange' },
-                    { name: 'Grapefruit', details: { color: 'Orangey' } }
-                ],
-                showSearch: true,
-                searchFields: ['name']
-            },
-            global: {
-                components: {
-                    FfDataTableRow, FfDataTableCell, FfTextInput, FfCheck, FfKebabMenu
-                }
-            }
-        })
-
-        expect(wrapper.find('tbody').text()).toBe('AppleOrangeGrapefruit')
-
-        await wrapper.find('[data-form="search"] input').setValue('Orange')
-
-        expect(wrapper.find('tbody').text()).toBe('Orange')
-    })
-})
-
 describe('DataTable', () => {
-    describe('filterRows', () => {
+    describe('#filterRows', () => {
         it('searches all properties and returns matching subset of rows', () => {
             const rows = [{
                 name: 'Apple',
@@ -409,7 +343,7 @@ describe('DataTable', () => {
         })
     })
 
-    describe('filteredRows', () => {
+    describe('#filteredRows', () => {
         it('Sorts numbers descending', () => {
             const localThis = {
                 lookupProperty: DataTable.methods.lookupProperty,
